requirements.txt
pyautogui==0.9.54
pandas==2.2.2
groq==0.9.0
openpyxl==3.1.2
python-dotenv==1.0.1
Pillow==10.3.0
